# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Welfare Settlement Exception Resolution Platform API
  description: |
    公益资金结算异常处置平台后端 API. 提供 项目/批次/周期/规则/明细/异常/汇总/工作台/审计 的 CRUD 与状态机操作.
    所有需要鉴权的接口期望 Bearer Token, 通过 X-Tenant-ID 指定租户 (默认 "default").
  version: 1.0.0
  contact:
    name: Welfare Settlement Resolver
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: http://127.0.0.1:8080/api/v1
    description: Local loopback
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  parameters:
    TenantID:
      name: X-Tenant-ID
      in: header
      description: 租户 ID (默认 "default")
      schema:
        type: string
        default: default
      required: false
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          example: INVALID_ARGUMENT
        message:
          type: string
        fields:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        request_id:
          type: string
    FieldError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    Exception:
      type: object
      properties:
        id: { type: string }
        tenant_id: { type: string }
        cycle_id: { type: string }
        entry_id: { type: string }
        rule_version_id: { type: string }
        rule_code: { type: string }
        severity: { type: string }
        title: { type: string }
        description: { type: string }
        hit_reason: { type: string }
        status: { type: string, enum: [pending, processing, review, resolved, closed, escalated] }
        assignee_id: { type: string, nullable: true }
        reporter_id: { type: string, nullable: true }
        deadline_at: { type: string, format: date-time, nullable: true }
        resolved_at: { type: string, format: date-time, nullable: true }
        closed_at: { type: string, format: date-time, nullable: true }
        version: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Summary:
      type: object
      properties:
        id: { type: string }
        tenant_id: { type: string }
        cycle_id: { type: string }
        rule_version_id: { type: string }
        computed_at: { type: string, format: date-time }
        total_entries: { type: integer }
        total_amount_cents: { type: integer, format: int64 }
        approved_amount_cents: { type: integer, format: int64 }
        pending_amount_cents: { type: integer, format: int64 }
        exception_count_by_status: { type: object, additionalProperties: { type: integer } }
        exception_count_by_severity: { type: object, additionalProperties: { type: integer } }
        version: { type: integer }
    AuditEntry:
      type: object
      properties:
        id: { type: string }
        tenant_id: { type: string }
        actor_id: { type: string }
        action: { type: string }
        entity_type: { type: string }
        entity_id: { type: string }
        detail: { type: object, additionalProperties: { type: string } }
        created_at: { type: string, format: date-time }
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
security:
  - bearerAuth: []
...
paths:
...
  /summaries/recalculate:
    parameters: [ { $ref: '#/components/parameters/TenantID' } ]
    post:
      tags: [Summaries]
      summary: 重新计算汇总 (保留快照)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cycle_id, rule_version_id, trigger_reason]
              properties:
                cycle_id: { type: string }
                rule_version_id: { type: string }
                trigger_reason: { type: string }
      responses:
        '201':
          description: 重新计算结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  recalc_id: { type: string }
                  summary: { $ref: '#/components/schemas/Summary' }
                  previous: { $ref: '#/components/schemas/Summary' }
  /summaries/cycles/{cycleId}/latest:
    parameters:
      - $ref: '#/components/parameters/TenantID'
      - name: cycleId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Summaries]
      summary: 查询周期最新汇总
      responses:
        '200':
          description: 汇总
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Summary' }
        '404': { $ref: '#/components/responses/NotFound' }
  /summaries/cycles/{cycleId}/history:
    parameters:
      - $ref: '#/components/parameters/TenantID'
      - name: cycleId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Summaries]
      summary: 查询周期汇总历史
      responses:
        '200':
          description: 历史
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Summary' } }
  /summaries/recalcs:
    parameters: [ { $ref: '#/components/parameters/TenantID' } ]
    get:
      tags: [Summaries]
      summary: 重算批次列表
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        cycle_id: { type: string }
                        rule_version_id: { type: string }
                        input_digest: { type: string }
                        trigger_reason: { type: string }
                        status: { type: string }
                        started_at: { type: string, format: date-time }
                        finished_at: { type: string, format: date-time, nullable: true }
                  page: { type: integer }
                  page_size: { type: integer }
                  total: { type: integer }
                  has_next: { type: boolean }
  /annual/{projectId}/{year}:
    parameters:
      - $ref: '#/components/parameters/TenantID'
      - name: projectId
        in: path
        required: true
        schema: { type: string }
      - name: year
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [Summaries]
      summary: 查询年度预算累计
      responses:
        '200':
          description: 年度累计
          content:
            application/json:
              schema:
                type: object
                properties:
                  project_id: { type: string }
                  year: { type: integer }
                  budget_cents: { type: integer, format: int64 }
                  disbursed_cents: { type: integer, format: int64 }
                  available_cents: { type: integer, format: int64 }
                  overrun_cents: { type: integer, format: int64 }
                  adjustments:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        delta_cents: { type: integer, format: int64 }
                        reason: { type: string }
                        triggered_by: { type: string }
                        created_at: { type: string, format: date-time }
  /annual/{projectId}/{year}/adjustments:
    parameters:
      - $ref: '#/components/parameters/TenantID'
      - name: projectId
        in: path
        required: true
        schema: { type: string }
      - name: year
        in: path
        required: true
        schema: { type: integer }
    post:
      tags: [Summaries]
      summary: 调整年度预算
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                delta_cents: { type: integer, format: int64 }
                reason: { type: string }
      responses:
        '200':
          description: 调整后年度累计
          content:
            application/json:
              schema:
                type: object
                properties:
                  project_id: { type: string }
                  year: { type: integer }
                  budget_cents: { type: integer, format: int64 }
                  disbursed_cents: { type: integer, format: int64 }
                  available_cents: { type: integer, format: int64 }
                  overrun_cents: { type: integer, format: int64 }
  /workspace/{assigneeId}:
    parameters:
      - $ref: '#/components/parameters/TenantID'
      - name: assigneeId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Workspace]
      summary: 查询处理人工作台
      responses:
        '200':
          description: 工作台
          content:
            application/json:
              schema:
                type: object
                properties:
                  assignee_id: { type: string }
                  open: { type: array, items: { $ref: '#/components/schemas/Exception' } }
                  overdue: { type: array, items: { $ref: '#/components/schemas/Exception' } }
                  escalated: { type: array, items: { $ref: '#/components/schemas/Exception' } }
                  recently_closed: { type: array, items: { $ref: '#/components/schemas/Exception' } }
  /audit:
    parameters: [ { $ref: '#/components/parameters/TenantID' } ]
    get:
      tags: [Audit]
      summary: 审计日志列表
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: actor_id
          in: query
          schema: { type: string }
        - name: action
          in: query
          schema: { type: string }
        - name: entity_id
          in: query
          schema: { type: string }
      responses:
        '200':
          description: 列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/AuditEntry' } }
                  page: { type: integer }
                  page_size: { type: integer }
                  total: { type: integer }
                  has_next: { type: boolean }
  /audit/export:
    parameters: [ { $ref: '#/components/parameters/TenantID' } ]
    get:
      tags: [Audit]
      summary: 导出审计日志 CSV
      responses:
        '200':
          description: CSV
          content:
            text/csv:
              schema: { type: string }
  /audit/exceptions/{cycleId}/export:
    parameters:
      - $ref: '#/components/parameters/TenantID'
      - name: cycleId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Audit]
      summary: 导出周期异常 CSV
      responses:
        '200':
          description: CSV
          content:
            text/csv:
              schema: { type: string }
  /users:
    parameters: [ { $ref: '#/components/parameters/TenantID' } ]
    post:
      tags: [Users]
      summary: 创建用户
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, role]
              properties:
                username: { type: string }
                display_name: { type: string }
                email: { type: string }
                role: { type: string, enum: [operator, assignee, reviewer, admin] }
                password_hash: { type: string }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  username: { type: string }
                  display_name: { type: string }
                  email: { type: string }
                  role: { type: string }
    get:
      tags: [Users]
      summary: 用户列表
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        username: { type: string }
                        display_name: { type: string }
                        email: { type: string }
                        role: { type: string }
                  page: { type: integer }
                  page_size: { type: integer }
                  total: { type: integer }
                  has_next: { type: boolean }